+++ prac_07/convert_miles_km.kv
@@ -3,24 +3,27 @@ BoxLayout:
     BoxLayout:
         orientation: 'horizontal'
         TextInput:
-            id: input_miles
+            id: input_number
             text: '34'
             font_size: 48
-            on_text: app.handle_calculate()
+            multiline: False
+            on_text: app.handle_calculate(input_number.text)
         BoxLayout:
             orientation: 'vertical'
-            size_hint_x: 0.25
             Button:
                 text: 'Up'
+                font_size: 30
                 on_press: app.handle_increment(1)
             Button:
                 text: 'Down'
+                font_size: 30
                 on_press: app.handle_increment(-1)
 #    Button:
-#        text: 'Convert m to km'
-#        on_press: app.handle_calculate()
+#        text: 'Convert miles to km'
+#        font_size: 40
+#        on_press: app.handle_calculate(input_number.text)
     Label:
         id: output_label
-        font_size: 48
         text: '54.717'
+        font_size: 48
         color: (1, 1, 0, 1)
